Rename utils.py -> io.py, operations.py -> op_utils.py
