+++ stac_mjx/__init__.py
@@ -1,5 +1,7 @@
 """This module exposes all high level APIs for stac-mjx."""
 
-from stac_mjx.utils import enable_xla_flags, load_data
+from stac_mjx.op_utils import enable_xla_flags
+from stac_mjx.io import load_data
 from stac_mjx.main import load_configs, run_stac
 from stac_mjx.viz import viz_stac
+
