+++ stac_mjx/compute_stac.py
@@ -7,7 +7,7 @@ from typing import Tuple, List
 import time
 
 from stac_mjx import stac_core
-from stac_mjx import operations as op
+from stac_mjx import op_utils
 
 
 def root_optimization(
@@ -72,7 +72,7 @@ def root_optimization(
 
     r = time.time()
 
-    mjx_data = op.replace_qs(mjx_model, mjx_data, op.make_qs(q0, qs_to_opt, res.params))
+    mjx_data = op_utils.replace_qs(mjx_model, mjx_data, op_utils.make_qs(q0, qs_to_opt, res.params))
     print(f"Replace 1 finished in {time.time()-r}")
 
     q0 = jp.copy(mjx_data.qpos[:])
@@ -96,7 +96,7 @@ def root_optimization(
     print(f"q_opt 2 finished in {time.time()-j} with an error of {res.state.error}")
     r = time.time()
 
-    mjx_data = op.replace_qs(mjx_model, mjx_data, op.make_qs(q0, qs_to_opt, res.params))
+    mjx_data = op_utils.replace_qs(mjx_model, mjx_data, op_utils.make_qs(q0, qs_to_opt, res.params))
 
     print(f"Replace 2 finished in {time.time()-r}")
     print(f"Root optimization finished in {time.time()-s}")
@@ -142,7 +142,7 @@ def offset_optimization(
     print("Begining offset optimization:")
 
     # Define initial position of the optimization
-    offset0 = op.get_site_pos(mjx_model, site_idxs).flatten()
+    offset0 = op_utils.get_site_pos(mjx_model, site_idxs).flatten()
 
     keypoints = jp.array(kp_data[time_indices, :])
     q = jp.take(q, time_indices, axis=0)
@@ -163,12 +163,12 @@ def offset_optimization(
     print(f"Final error of {res.state.error}")
 
     # Set body sites according to optimized offsets
-    mjx_model = op.set_site_pos(
+    mjx_model = op_utils.set_site_pos(
         mjx_model, jp.reshape(offset_opt_param, (-1, 3)), site_idxs
     )
 
     # Forward kinematics, and save the results to the walker sites as well
-    mjx_data = op.kinematics(mjx_model, mjx_data)
+    mjx_data = op_utils.kinematics(mjx_model, mjx_data)
 
     print(f"offset optimization finished in {time.time()-s}")
 
@@ -226,7 +226,7 @@ def pose_optimization(
             site_idxs,
         )
 
-        mjx_data = op.replace_qs(mjx_model, mjx_data, res.params)
+        mjx_data = op_utils.replace_qs(mjx_model, mjx_data, res.params)
 
         for part in parts:
             q0 = jp.copy(mjx_data.qpos[:])
@@ -243,8 +243,8 @@ def pose_optimization(
                 site_idxs,
             )
 
-            mjx_data = op.replace_qs(
-                mjx_model, mjx_data, op.make_qs(q0, part, res.params)
+            mjx_data = op_utils.replace_qs(
+                mjx_model, mjx_data, op_utils.make_qs(q0, part, res.params)
             )
 
         return mjx_data, res.state.error
@@ -259,7 +259,7 @@ def pose_optimization(
 
         q.append(mjx_data.qpos[:])
         x.append(mjx_data.xpos[:])
-        walker_body_sites.append(op.get_site_xpos(mjx_data, site_idxs))
+        walker_body_sites.append(op_utils.get_site_xpos(mjx_data, site_idxs))
 
         frame_time.append(time.time() - loop_start)
         frame_error.append(error)
