+++ stac_mjx/op_utils.py
@@ -2,11 +2,19 @@
 
 from jax import numpy as jp
 from jax import jit
+from jax.lib import xla_bridge
 from mujoco import mjx
 from mujoco.mjx._src import smooth
 import numpy as np
-from stac_mjx import utils
+from stac_mjx import io
 
+def enable_xla_flags():
+    """Enables XLA Flags for faster runtime on Nvidia GPUs."""
+    if xla_bridge.get_backend().platform == "gpu":
+        os.environ["XLA_FLAGS"] = (
+            "--xla_gpu_enable_triton_softmax_fusion=true "
+            "--xla_gpu_triton_gemm_any=True "
+        )
 
 @jit
 def kinematics(mjx_model: mjx.Model, mjx_data: mjx.Data):
