+++ tests/test_controller.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
 from stac_mjx import main
-from stac_mjx import utils
+from stac_mjx import io
 from stac_mjx.stac import Stac, _align_joint_dims
 from mujoco import _structs
 
@@ -11,7 +11,7 @@ _BASE_PATH = Path.cwd()
 def test_init_stac(mocap_nwb, config):
     cfg = main.load_configs(config)
     xml_path = _BASE_PATH / cfg.model.MJCF_PATH
-    kp_data, sorted_kp_names = utils.load_data(cfg)
+    kp_data, sorted_kp_names = io.load_data(cfg)
 
     stac = Stac(xml_path, cfg, sorted_kp_names)
