+++ tests/test_io.py
@@ -1,4 +1,4 @@
-from stac_mjx import utils
+from stac_mjx import io
 from pathlib import Path
 import pytest
 import hydra
@@ -31,7 +31,7 @@ def test_load_nwb(config, mocap_nwb):
     # params = utils._load_params(_BASE_PATH / rodent_config)
     # assert params is not None
     cfg = load_config_with_overrides(config, stac_data_path_override=mocap_nwb)
-    data, sorted_kp_names = utils.load_data(cfg)
+    data, sorted_kp_names = io.load_data(cfg)
     assert data.shape == (1000, 69)
     assert len(sorted_kp_names) == 23
 
@@ -41,7 +41,7 @@ def test_load_mat_no_label3d(config, mocap_mat):
     Test loading data from .mat file.
     """
     cfg = load_config_with_overrides(config, stac_data_path_override=mocap_mat)
-    data, sorted_kp_names = utils.load_data(cfg)
+    data, sorted_kp_names = io.load_data(cfg)
     assert data.shape == (1000, 69)
     assert len(sorted_kp_names) == 23
 
@@ -53,7 +53,7 @@ def test_load_mat_w_label3d(config, rodent_config_label3d, mocap_mat):
     cfg = load_config_with_overrides(
         config, stac_data_path_override=mocap_mat, model_override=rodent_config_label3d
     )
-    data, sorted_kp_names = utils.load_data(cfg)
+    data, sorted_kp_names = io.load_data(cfg)
     assert data.shape == (1000, 69)
     assert len(sorted_kp_names) == 23
 
@@ -66,7 +66,7 @@ def test_load_mat_no_kp_names(config, rodent_config_no_kp_names, mocap_mat):
     )
 
     with pytest.raises(ValueError):
-        data, sorted_kp_names = utils.load_data(cfg)
+        data, sorted_kp_names = io.load_data(cfg)
 
 
 def test_load_mat_less_kp_names(config, rodent_config_less_kp_names, mocap_mat):
@@ -77,7 +77,7 @@ def test_load_mat_less_kp_names(config, rodent_config_less_kp_names, mocap_mat):
     )
 
     with pytest.raises(ValueError):
-        data, sorted_kp_names = utils.load_data(cfg)
+        data, sorted_kp_names = io.load_data(cfg)
 
 
 def test_load_h5(config, mouse_config, mocap_h5):
@@ -91,7 +91,7 @@ def test_load_h5(config, mouse_config, mocap_h5):
         model_override=mouse_config,
     )
 
-    data, sorted_kp_names = utils.load_data(cfg)
+    data, sorted_kp_names = io.load_data(cfg)
 
     assert data.shape == (3600, 102)
     assert len(sorted_kp_names) == 34
